backend/requirements.txt
fastapi
uvicorn
openai
langdetect
python-dotenv